+++ Platform/FO-DICOM.AspNetCore/Configs/ServerConfiguration.cs
@@ -1,0 +1,9 @@
+﻿namespace FellowOakDicom.AspNetCore.Configs
+{
+    public class ServerConfiguration
+    {
+        public int Port { get; set; } = 104;
+
+        public string AETitle { get; set; } = "FODICOMSERVER";
+    }
+}